+++ CHANGELOG.md
@@ -43,6 +43,8 @@ Release TBD
   * Deprecate updating cart currency, just purge carts if currency is changed.
   * Allow sites to define where each address field is required.
   * Combine enable-ratings configs for logged-in and anon users.
+  * Include SKU in item purchase report.
+  * Fix setting SKU in order item object to include the variant.
 
 ## v1.4.1
 Release 2022-04-07